backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.11
django-cors-headers==4.9.0
django-daraja==1.3.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
